handle dot expr head specially (#19)
Treat `.` expressions as syntax rather than equivalent to a
`getproperty` call. This is a minor breaking change if someone was
previously relying on it being treated as a call-expr, as they must now
rewrite that as an explicit call to `getfield`. However, I think, on
balance, that syntax is likely to be clearer at highlighting the user's
likely intent.